+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/DefaultListArbitrary.java
@@ -1,12 +1,15 @@
 package net.jqwik.engine.properties.arbitraries;
 
 import java.util.*;
+import java.util.function.*;
+import java.util.stream.*;
 
 import net.jqwik.api.*;
+import net.jqwik.api.arbitraries.*;
 import net.jqwik.engine.properties.arbitraries.exhaustive.*;
 import net.jqwik.engine.properties.shrinking.*;
 
-public class DefaultListArbitrary<T> extends MultivalueArbitraryBase<T, List<T>> {
+public class DefaultListArbitrary<T> extends MultivalueArbitraryBase<T, List<T>> implements ListArbitrary<T> {
 
 	public DefaultListArbitrary(Arbitrary<T> elementArbitrary, boolean elementsUnique) {
 		super(elementArbitrary, elementsUnique);
@@ -32,4 +35,38 @@ public class DefaultListArbitrary<T> extends MultivalueArbitraryBase<T, List<T>>
 		return edgeCases(ShrinkableList::new);
 	}
 
+	@Override
+	public ListArbitrary<T> ofMaxSize(int maxSize) {
+		return (ListArbitrary<T>) super.ofMaxSize(maxSize);
+	}
+
+	@Override
+	public ListArbitrary<T> ofMinSize(int minSize) {
+		return (ListArbitrary<T>) super.ofMinSize(minSize);
+	}
+
+	@Override
+	public ListArbitrary<T> ofSize(int size) {
+		return ofMinSize(size).ofMaxSize(size);
+	}
+
+	// TODO: Remove duplication with DefaultSetArbitrary.mapEach()
+	@Override
+	public <U> Arbitrary<List<U>> mapEach(BiFunction<List<T>, T, U> mapper) {
+		return this.map(elements -> elements.stream()
+											.map(e -> mapper.apply(elements, e))
+											.collect(Collectors.toList()));
+	}
+
+	// TODO: Remove duplication with DefaultSetArbitrary.flatMapEach()
+	@Override
+	public <U> Arbitrary<List<U>> flatMapEach(BiFunction<List<T>, T, Arbitrary<U>> flatMapper) {
+		return this.flatMap(elements -> {
+			List<Arbitrary<U>> arbitraries =
+				elements.stream()
+						.map(e -> flatMapper.apply(elements, e))
+						.collect(Collectors.toList());
+			return Combinators.combine(arbitraries).as(ArrayList::new);
+		});
+	}
 }