+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/DefaultSetArbitrary.java
@@ -42,29 +42,31 @@ public class DefaultSetArbitrary<T> extends MultivalueArbitraryBase<T, Set<T>> i
 	}
 
 	@Override
-	public SetArbitrary<T> ofMaxSize(final int maxSize) {
+	public SetArbitrary<T> ofMaxSize(int maxSize) {
 		return (SetArbitrary<T>) super.ofMaxSize(maxSize);
 	}
 
 	@Override
-	public SetArbitrary<T> ofMinSize(final int minSize) {
+	public SetArbitrary<T> ofMinSize(int minSize) {
 		return (SetArbitrary<T>) super.ofMinSize(minSize);
 	}
 
 	@Override
-	public SetArbitrary<T> ofSize(final int size) {
+	public SetArbitrary<T> ofSize(int size) {
 		return this.ofMinSize(size).ofMaxSize(size);
 	}
 
+	// TODO: Remove duplication with DefaultListArbitrary.mapEach()
 	@Override
-	public <U> Arbitrary<Set<U>> mapEach(final BiFunction<Set<T>, T, U> mapper) {
+	public <U> Arbitrary<Set<U>> mapEach(BiFunction<Set<T>, T, U> mapper) {
 		return this.map(elements -> elements.stream()
-									.map(e -> mapper.apply(elements, e))
-									.collect(Collectors.toSet()));
+											.map(e -> mapper.apply(elements, e))
+											.collect(Collectors.toSet()));
 	}
 
+	// TODO: Remove duplication with DefaultListArbitrary.flatMapEach()
 	@Override
-	public <U> Arbitrary<Set<U>> flatMapEach(final BiFunction<Set<T>, T, Arbitrary<U>> flatMapper) {
+	public <U> Arbitrary<Set<U>> flatMapEach(BiFunction<Set<T>, T, Arbitrary<U>> flatMapper) {
 		return this.flatMap(elements -> {
 			List<Arbitrary<U>> arbitraries =
 				elements.stream()